Lab jboren/gpuw 2018 06

+++ oneEngine/oneGame/source/audio/Listener.cpp
@@ -11,11 +11,11 @@ audio::Listener::Listener( void )
 	}
 
 	// Set initial options
-	position = Vector3d();
-	position_prev = Vector3d();
-	velocity = Vector3d();
-	orient_forward = Vector3d::forward;
-	orient_up = Vector3d( 0,0,1 );
+	position = Vector3f();
+	position_prev = Vector3f();
+	velocity = Vector3f();
+	orient_forward = Vector3f::forward;
+	orient_up = Vector3f( 0,0,1 );
 }
 
 audio::Listener::~Listener ( void )
