+++ oneEngine/oneGame/source/core/math/Cubic.cpp
@@ -7,17 +7,17 @@ using namespace core::math;
 // Default empty constructor
 Cubic::Cubic ( void )
 {
-	position = Vector3d( 0,0,0 );
-	size = Vector3d( 1,1,1 );
+	position = Vector3f( 0,0,0 );
+	size = Vector3f( 1,1,1 );
 }
 // Constructor taking two vectors
-Cubic::Cubic ( Vector3d const& vPosition, Vector3d const& vSize )
+Cubic::Cubic ( Vector3f const& vPosition, Vector3f const& vSize )
 {
 	position = vPosition;
 	size = vSize;
 }
 // Creation taking min and max pos
-Cubic Cubic::FromPosition ( Vector3d vMinPos, Vector3d vMaxPos )
+Cubic Cubic::FromPosition ( Vector3f vMinPos, Vector3f vMaxPos )
 {
 	if ( vMinPos.x > vMaxPos.x ) {
 		Real temp = vMinPos.x;
@@ -56,26 +56,26 @@ void Cubic::Realign( void )
 }
 
 // Get center
-Vector3d Cubic::center ( void ) const
+Vector3f Cubic::center ( void ) const
 {
 	return ( position + (size*0.5f) );
 }
 
 // Returns a frustum
-Frustum Cubic::GetFrustum ( Vector3d const& casterPosition )
+Frustum Cubic::GetFrustum ( Vector3f const& casterPosition )
 {
 	// 0,0,0,bottom,right,backward
 	// First gotta choose what to generate the frustum with.
 	// This shouldn't be too hard. Based on the closest vertex to the camera, choose the other 6 vertices to use.
-	Vector3d places [8] = {
+	Vector3f places [8] = {
 		position,
-		position + Vector3d( size.x,0,0 ),
-		position + Vector3d( 0,size.y,0 ),
-		position + Vector3d( size.x,size.y,0 ),
-		position + Vector3d( 0,0,size.z ),
-		position + Vector3d( size.x,0,size.z ),
-		position + Vector3d( 0,size.y,size.z ),
-		position + Vector3d( size.x,size.y,size.z )
+		position + Vector3f( size.x,0,0 ),
+		position + Vector3f( 0,size.y,0 ),
+		position + Vector3f( size.x,size.y,0 ),
+		position + Vector3f( 0,0,size.z ),
+		position + Vector3f( size.x,0,size.z ),
+		position + Vector3f( 0,size.y,size.z ),
+		position + Vector3f( size.x,size.y,size.z )
 	};
 	const static int cube_lookup [8][6] = {
 		{3,1,5,4,6,2},
@@ -114,7 +114,7 @@ Frustum Cubic::GetFrustum ( Vector3d const& casterPosition )
 }
 
 // Check if point is inside the box
-bool Cubic::PointIsInside ( Vector3d const& point )
+bool Cubic::PointIsInside ( Vector3f const& point )
 {
 	if ( point.x < position.x ) return false;
 	if ( point.y < position.y ) return false;
@@ -124,28 +124,28 @@ bool Cubic::PointIsInside ( Vector3d const& point )
 	if ( point.z > position.z+size.z ) return false;
 	return true;
 }
-/*#include "../CDebugDrawer.h"
+/*#include "../RrDebugDrawer.h"
 void Cubic::DrawDebug ( void )
 {
-	debug::Drawer->DrawLine( position+Vector3d( 0,0,0 ), position+Vector3d( size.x,0,0 ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,0,0 ), position+Vector3d( size.x,size.y,0 ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,size.y,0 ), position+Vector3d( 0,size.y,0 ) );
-	debug::Drawer->DrawLine( position+Vector3d( 0,size.y,0 ), position+Vector3d( 0,0,0 ) );
-	debug::Drawer->DrawLine( position+Vector3d( 0,0,size.z ), position+Vector3d( size.x,0,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,0,size.z ), position+Vector3d( size.x,size.y,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,size.y,size.z ), position+Vector3d( 0,size.y,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( 0,size.y,size.z ), position+Vector3d( 0,0,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( 0,0,0 ), position+Vector3d( 0,0,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,0,0 ), position+Vector3d( size.x,0,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( size.x,size.y,0 ), position+Vector3d( size.x,size.y,size.z ) );
-	debug::Drawer->DrawLine( position+Vector3d( 0,size.y,0 ), position+Vector3d( 0,size.y,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,0,0 ), position+Vector3f( size.x,0,0 ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,0,0 ), position+Vector3f( size.x,size.y,0 ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,size.y,0 ), position+Vector3f( 0,size.y,0 ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,size.y,0 ), position+Vector3f( 0,0,0 ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,0,size.z ), position+Vector3f( size.x,0,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,0,size.z ), position+Vector3f( size.x,size.y,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,size.y,size.z ), position+Vector3f( 0,size.y,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,size.y,size.z ), position+Vector3f( 0,0,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,0,0 ), position+Vector3f( 0,0,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,0,0 ), position+Vector3f( size.x,0,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( size.x,size.y,0 ), position+Vector3f( size.x,size.y,size.z ) );
+	debug::Drawer->DrawLine( position+Vector3f( 0,size.y,0 ), position+Vector3f( 0,size.y,size.z ) );
 }*/
 
 // Check if box intersects with box
 bool Cubic::CubicCollides ( const Cubic & cube )
 {
-	Vector3d cubePosition = (position+size*0.5f)-(cube.position+cube.size*0.5f);
-	Vector3d cubeExtents = (size + cube.size)*0.5f;
+	Vector3f cubePosition = (position+size*0.5f)-(cube.position+cube.size*0.5f);
+	Vector3f cubeExtents = (size + cube.size)*0.5f;
 	
 	if ( fabs( cubePosition.x ) > cubeExtents.x ) return false;
 	if ( fabs( cubePosition.y ) > cubeExtents.y ) return false;
@@ -157,7 +157,7 @@ bool Cubic::CubicCollides ( const Cubic & cube )
 // Check if line intersects with the box
 bool Cubic::LineCollides ( Line const& line )
 {
-	/*Vector3d position2 = position+size;
+	/*Vector3f position2 = position+size;
 
 	if (line.end.x < position.x  && line.start.x < position.x) return false;
 	if (line.end.x > position2.x && line.start.x > position2.x) return false;
@@ -171,12 +171,12 @@ bool Cubic::LineCollides ( Line const& line )
 
 	return false;*/
 
-	Vector3d cubeCenter = position+size*Real(0.5);
-	Vector3d cubeExtents = size*Real(0.5);
+	Vector3f cubeCenter = position+size*Real(0.5);
+	Vector3f cubeExtents = size*Real(0.5);
 
-	Vector3d lineMid = ((line.midpoint())-cubeCenter);
-	Vector3d lineDir = ((line.start-cubeCenter) - lineMid);
-	Vector3d lineExt = Vector3d( fabs(lineDir.x), fabs(lineDir.y), fabs(lineDir.z) );
+	Vector3f lineMid = ((line.midpoint())-cubeCenter);
+	Vector3f lineDir = ((line.start-cubeCenter) - lineMid);
+	Vector3f lineExt = Vector3f( fabs(lineDir.x), fabs(lineDir.y), fabs(lineDir.z) );
 
 	// Use Separating Axis Test
 	// Separation vector from box center to line center is LMid, since the line is in box space
@@ -194,9 +194,9 @@ bool Cubic::LineCollides ( Line const& line )
 
 //http://www.3dkingdoms.com/weekly/weekly.php?a=3
 // Check if line intersects with the box. Get the first intersection point
-/*bool CCubic::LineGetCollision ( Line const& line, Vector3d & outPos )
+/*bool CCubic::LineGetCollision ( Line const& line, Vector3f & outPos )
 {
-	Vector3d position2 = position+size;
+	Vector3f position2 = position+size;
 
 	if (line.end.x < position.x  && line.start.x < position.x) return false;
 	if (line.end.x > position2.x && line.start.x > position2.x) return false;
@@ -223,7 +223,7 @@ bool Cubic::LineCollides ( Line const& line )
 }
 
 // Helper functions for collision
-bool inline CCubic::GetIntersection( Real fDst1, Real fDst2, Vector3d P1, Vector3d P2, Vector3d & Hit )
+bool inline CCubic::GetIntersection( Real fDst1, Real fDst2, Vector3f P1, Vector3f P2, Vector3f & Hit )
 {
 	if ( (fDst1 * fDst2) >= 0.0f) return false;
 	//if ( fDst1 == fDst2 ) return false; 
@@ -231,7 +231,7 @@ bool inline CCubic::GetIntersection( Real fDst1, Real fDst2, Vector3d P1, Vector
 	Hit = P1 + (P2-P1) * ( -fDst1/(fDst2-fDst1) );
 	return true;
 }
-bool inline CCubic::InBox( Vector3d Hit, Vector3d B1, Vector3d B2, const char Axis )
+bool inline CCubic::InBox( Vector3f Hit, Vector3f B1, Vector3f B2, const char Axis )
 {
 	if ( Axis==1 && Hit.z > B1.z && Hit.z < B2.z && Hit.y > B1.y && Hit.y < B2.y) return true;
 	if ( Axis==2 && Hit.z > B1.z && Hit.z < B2.z && Hit.x > B1.x && Hit.x < B2.x) return true;
@@ -264,7 +264,7 @@ bool rayAABBIntersect1D(float start, float dir, float min, float max, float& ent
     if(t1 < exit) exit = t1;
     return true;
 }
-bool rayAABBIntersect(Vector3d start, Vector3d dir, Vector3d min, Vector3d max, Vector3d& penter, Vector3d& pexit)
+bool rayAABBIntersect(Vector3f start, Vector3f dir, Vector3f min, Vector3f max, Vector3f& penter, Vector3f& pexit)
 {
     float enter = 0.0f, exit = 1.0f;
 
@@ -281,8 +281,8 @@ bool rayAABBIntersect(Vector3d start, Vector3d dir, Vector3d min, Vector3d max, 
     pexit  = start + dir * exit;
     return true;
 }
-bool Cubic::LineGetCollision ( Line const& line, Vector3d & outPos )
+bool Cubic::LineGetCollision ( Line const& line, Vector3f & outPos )
 {
-	Vector3d bleh;
+	Vector3f bleh;
 	return rayAABBIntersect( line.start, line.end-line.start, position, position+size, outPos, bleh );
 }
