+++ oneEngine/oneGame/source/core/math/Cubic.h
@@ -5,7 +5,7 @@
 #define CORE_MATH_CUBIC_H_
 
 // Includes
-#include "Vector3d.h"
+#include "Vector3.h"
 #include "Line.h"
 #include "Frustum.h"
 
@@ -18,37 +18,37 @@ namespace core
 		{
 		public:
 			// Bottom back right corner
-			Vector3d position;
+			Vector3f position;
 			// Size of object, going forward x, left y, and up z
-			Vector3d size;
+			Vector3f size;
 
 		public:
 			Cubic ( void );
-			Cubic ( Vector3d const& vPosition, Vector3d const& vSize );
+			Cubic ( Vector3f const& vPosition, Vector3f const& vSize );
 
 			//		FromPosition ( pos1, pos2 )
 			// Creates a new Cubic from the input positions
-			static Cubic FromPosition ( Vector3d vMinPos, Vector3d vMaxPos );
+			static Cubic FromPosition ( Vector3f vMinPos, Vector3f vMaxPos );
 
 			//		Realign ()
 			// Looks at the position and size, and moves their values around so that
 			// the position refers to the bottom back right corner of the Cubic.
 			void Realign ( void );
 
-			Vector3d center ( void ) const;
+			Vector3f center ( void ) const;
 		public:
 			void DrawDebug ( void );
 		public:
-			Frustum GetFrustum ( Vector3d const& casterPosition );
-			bool PointIsInside ( Vector3d const& );
+			Frustum GetFrustum ( Vector3f const& casterPosition );
+			bool PointIsInside ( Vector3f const& );
 			bool LineCollides ( Line const& );
-			bool LineGetCollision ( Line const&, Vector3d & );
+			bool LineGetCollision ( Line const&, Vector3f & );
 	
 			bool CubicCollides ( const Cubic & );
 		private:
 			// Helper functions for collision
-			bool inline GetIntersection( Real fDst1, Real fDst2, Vector3d P1, Vector3d P2, Vector3d & Hit );
-			bool inline InBox( Vector3d Hit, Vector3d B1, Vector3d B2, const char Axis );
+			bool inline GetIntersection( Real fDst1, Real fDst2, Vector3f P1, Vector3f P2, Vector3f & Hit );
+			bool inline InBox( Vector3f Hit, Vector3f B1, Vector3f B2, const char Axis );
 		};
 	}
 }
