+++ oneEngine/oneGame/source/core/math/Ray.h
@@ -4,17 +4,17 @@
 #define _STRUCT_RAY_H_
 
 // Includes
-#include "Vector3d.h"
+#include "Vector3.h"
 
 // Struct Definition
 struct Ray
 {
 public:
 	Ray ( void ) : pos(0,0,0), dir(0,0,0) {;}
-	Ray ( const Vector3d & npos, const Vector3d & ndir ) : pos(npos), dir(ndir) {;}
+	Ray ( const Vector3f & npos, const Vector3f & ndir ) : pos(npos), dir(ndir) {;}
 public:
-	Vector3d pos;
-	Vector3d dir;
+	Vector3f pos;
+	Vector3f dir;
 };
 
 #endif
